requirements.txt
black==25.1.0
blinker==1.9.0
click==8.2.0
Deprecated==1.2.18
Flask==3.1.1
Flask-Limiter==3.12
flask-talisman==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
ordered-set==4.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
Pygments==2.19.1
pypdf==5.5.0
python-dotenv==1.1.0
python-magic==0.4.27
rich==13.9.4
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
